requirements.txt
openai==0.27.0
gradio==2.2.6
youtube-search-python==1.4.2
